pyproject.toml
[project]
name = "securetask"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.7",
  "uvicorn[standard]==0.34.0",
  "SQLAlchemy==2.0.36",
  "alembic==1.14.0",
  "psycopg[binary]==3.2.3",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "pydantic==2.10.3",
  "pydantic-settings==2.6.1",
  "python-multipart==0.0.20"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
